i'm an idiot
Rename vintageex subpackage to vex so that setup.py spa won't clean it up
(because it's the temp dir it uses to package the .sublime-text package).
This is a consequence of the case-insensitivitiy of the Windows file system
(its case-preserving nature also contributed to the confusion).

--HG--
rename : vintageex/parsing.py => vex/parsing.py